autoPOD/requirements.txt
# AI / LLM
anthropic>=0.25.0
openai>=1.30.0

# Scheduling
apscheduler>=3.10.4

# Database
sqlalchemy>=2.0.30
alembic>=1.13.1

# Configuration
python-dotenv>=1.0.1

# HTTP / Scraping
requests>=2.31.0
aiohttp>=3.9.5
beautifulsoup4>=4.12.3
selenium>=4.20.0
playwright>=1.44.0
undetected-chromedriver>=3.5.5
pytrends>=4.9.2

# Image processing
Pillow>=10.3.0
rembg>=2.0.57

# Data & reporting
pandas>=2.2.2
matplotlib>=3.9.0
plotly>=5.22.0
jinja2>=3.1.4

# Platform clients
praw>=7.7.1
slack-sdk>=3.27.2
sendgrid>=6.11.0

# Utilities
retry>=0.9.2
tenacity>=8.3.0
